Add migration for bulk_actions.body being mediumtext on mysql
Text field has a length of 65,535 characters which can easily be reached
if triggering a backfill with lots of partitions.

This changes the column type to mediumtext which has a character limit
of 16,777,215 characters which should be ample (don't think we need to
go to longtext just yet).

Sqlite and Postgres do not have a length limit on text, so this will
only apply to Mysql.